requirements.txt
Pyrogram
TgCrypto
pyromod
python-dotenv
dnspython
motor
Pillow
hachoir
mutagen
psutil
aiofiles
